+++ test/modules/core/controllers/index.ts
@@ -5,4 +5,5 @@
 import './controllers.spec';
 import './custom-controller.spec';
 import './globe-navigation.spec';
+import './globe-rubber-band.spec';
 import './view-states.spec';